Move Layout and components all to same directory
