+++ app/layout.tsx
@@ -1,7 +1,9 @@
 import '@prezly/content-renderer-react-js/styles.css';
 import '@prezly/uploadcare-image/build/styles.css';
 import '../styles/styles.globals.scss';
-import { Layout } from '@/components/Layout';
+
+import { Footer } from '@/components/Footer';
+import { Header } from '@/components/Header';
 
 import { Providers } from './providers';
 
@@ -35,7 +37,16 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
             <body className="flex h-full bg-zinc-50 dark:bg-black">
                 <Providers>
                     <div className="flex w-full">
-                        <Layout>{children}</Layout>
+                        <div className="fixed inset-0 flex justify-center sm:px-8">
+                            <div className="flex w-full max-w-7xl lg:px-8">
+                                <div className="w-full bg-white ring-1 ring-zinc-100 dark:bg-zinc-900 dark:ring-zinc-300/20" />
+                            </div>
+                        </div>
+                        <div className="relative flex w-full flex-col">
+                            <Header />
+                            <main className="flex-auto">{children}</main>
+                            <Footer />
+                        </div>
                     </div>
                 </Providers>
             </body>
