+++ components/SubscribeForm/MailIcon.tsx
