--- components/types.ts
@@ -1,1 +1,0 @@
-export type HeadingTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
