--- hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useDevice';
-export * from './useIsMounted';
