+++ lib/hooks/useDevice.ts
