+++ tsconfig.json
@@ -28,17 +28,8 @@
       "@/contexts/*": [
         "contexts/*"
       ],
-      "@/hooks": [
-        "hooks/"
-      ],
-      "@/hooks/*": [
-        "hooks/*"
-      ],
       "@/icons": [
         "icons/"
-      ],
-      "@/modules/*": [
-        "modules/*"
       ],
       "@/public/*": [
         "public/*"
